Inject an init function that can be used to initialize plugins.
An heuristic is used to find the possible entry points in the lifted program:
if a function is never the explicit target of a CALL pcode, then it could be an entry point.
Even though the function can be called multiple times (if a function is mistakenly considered an entry point), its body will be executed only once.